Rework play/stop button UI
We no longer change the text of the button (play -> stop).

Instead, the "play" button is now a checkbox element with a label styled
like a button element. When it's checked, we style the label like a
pressed button until autoplay stops.

Also, we disable the "next" button when autoplay is on.

This is how Verou implemented the autoplay checkbox/button in the
course, and I took her CSS (button shadows) like before.

She used the clip CSS property to hide checkbox-play while keeping
accessibility [1], but it seems that property is deprecated. So I used
the newer clip-path [2] just in case. I don't really know if this
remains accessible to screen readers though...

[1] https://snook.ca/archives/html_and_css/hiding-content-for-accessibility
[2] https://developer.mozilla.org/en-US/docs/Web/CSS/clip-path